requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
gunicorn
djangorestframework-simplejwt
drf_spectacular
django-jet-reboot
django-cors-headers
requests